Refactor test configuration and structure: - Updated ESLint configuration to include e2e tests. - Changed Playwright test directory from './tests' to './e2e'. - Adjusted TypeScript configuration to include e2e files. - Removed obsolete test files: example.spec.ts and wallet.spec.ts.
Closes #243
