+++ tsconfig.test.json
@@ -1,6 +1,6 @@
 {
   "extends": "./tsconfig.json",
-  "include": ["tests/**/*", "playwright.config.ts"],
+  "include": ["e2e/**/*", "playwright.config.ts"],
   "compilerOptions": {
     "types": ["@playwright/test"]
   }
